refactor: complete Phase 2 file migration
