+++ src/main.jsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import App from './App.jsx'
-import { ToastProvider } from './components/Toast.jsx'
+import App from './App'
+import { ToastProvider } from './components/ui/Toast'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
